# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cairn API",
    "description": "\nCairn walks a family through the logistics of a death, one step at a time.\n\nEvery response that moves a conversation forward carries a `next_step` with a\nsingle prompt, so clients can ask one question at a time. Errors use RFC 9457\nproblem details (`application/problem+json`) and never echo submitted values.\n\nAuthentication is an Auth0 access token. Accounts can be created with Google,\nApple, or an email magic link (see GET /v1/welcome). Onboarding follows\nUC-REG-01 to UC-REG-14, one question per screen. Case creation follows\nUC-CASE-01 to UC-CASE-18: a case starts as a draft, and the free period of 28\ndays starts only when the user taps Start journey on their first case. After\nit ends the account is read-only until subscribed. No payment information is\nasked for anywhere in case creation. Free text is redacted as it is received\nand never stored. The case is the security boundary and is enforced in the\ndatabase with row-level security.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/cases/{case_id}": {
      "get": {
        "tags": [
          "Case creation"
        ],
        "summary": "Open a case",
        "description": "UC-CASE-10. Opening a draft counts as activity and resets its 28 days. The response greets the user, says in one line where they left off, and asks whether to continue.",
        "operationId": "get_case_v1_cases__case_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not a member of this case, the case does not exist, or (for an active case) the account is read-only (code account_read_only)."
          },
          "409": {
            "description": "Onboarding isn't finished, or a changed acknowledgment needs to be agreed to."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/cases/{case_id}/journey/preview": {
      "get": {
        "tags": [
          "Journey"
        ],
        "summary": "See the journey that fits",
        "description": "UC-CASE-12. Runs journey selection on the answers so far and explains in a sentence or two why the journey fits. Shows the first four weeks by week: done items checked, open items labeled When you're ready. Missing answers never block it. Shows the pre-button notice, then Start journey and Not yet. Viewing it never starts the free period.",
        "operationId": "preview_v1_cases__case_id__journey_preview_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JourneyPreviewResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not a member of this case, not an owner, or the case does not exist."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cases/{case_id}/journey/start": {
      "post": {
        "tags": [
          "Journey"
        ],
        "summary": "Start journey",
        "description": "UC-CASE-12 and UC-CASE-18. In one transaction: the case becomes active, the journey template key and version are pinned, journey_started_at is set, the tasks are created, and on the account's first journey the 28-day free period starts (DEC-01). A second journey never restarts or extends it (DEC-04). No payment information is asked for or stored (DEC-02). Send back the pre-button notice version to show the notice was on screen first. A read-only account gets a subscription prompt.",
        "operationId": "start_v1_cases__case_id__journey_start_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartJourneyIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartJourneyResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not a member of this case, not an owner, or the case does not exist."
          },
          "409": {
            "description": "Already started, the death hasn't happened, or the notice changed since it was shown."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cases/{case_id}/journey/not-yet": {
      "post": {
        "tags": [
          "Journey"
        ],
        "summary": "Not yet",
        "description": "UC-CASE-12. The case stays a draft, the free period doesn't start, and everything is saved.",
        "operationId": "not_yet_v1_cases__case_id__journey_not_yet_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FirstTaskResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not a member of this case, not an owner, or the case does not exist."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cases/{case_id}/journey/first-task": {
      "post": {
        "tags": [
          "Journey"
        ],
        "summary": "Choose what feels doable right now",
        "description": "UC-CASE-13. A task id opens that task (any task, without comment), small_task offers the small thing, and not_today leaves the case active with no task in progress. Cairn never starts a task the user didn't choose.",
        "operationId": "first_task_v1_cases__case_id__journey_first_task_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FirstTaskIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FirstTaskResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not a member of this case, not an owner, or the case does not exist."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cases/{case_id}/journey": {
      "get": {
        "tags": [
          "Journey"
        ],
        "summary": "The journey, with one next action in front",
        "description": "UC-9 and UC-12. While paused, a calm check-in and no task list. A draft points to the preview.",
        "operationId": "get_journey_v1_cases__case_id__journey_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JourneyResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not a member of this case, not an owner, or the case does not exist."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cases/{case_id}/journey/pause": {
      "post": {
        "tags": [
          "Journey"
        ],
        "summary": "Step back from tasks for a while",
        "description": "UC-12. Holds all progress exactly as it is. Stores only the time the pause ends. No reason or feeling is collected or stored.",
        "operationId": "pause_journey_v1_cases__case_id__journey_pause_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/PauseRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Req"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JourneyResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not a member of this case, not an owner, or the case does not exist."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CaseResponse": {
        "properties": {
          "case": {
            "$ref": "#/components/schemas/CaseOut"
          },
          "deceased": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeceasedOut"
              },
              {
                "type": "null"
              }
            ],
            "description": "Only once a task has collected it."
          },
          "acknowledgment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acknowledgment"
          },
          "body": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Body"
          },
          "notes": {
            "items": {
              "$ref": "#/components/schemas/Note"
            },
            "type": "array",
            "title": "Notes"
          },
          "next_step": {
            "$ref": "#/components/schemas/NextStep"
          },
          "read_aloud": {
            "$ref": "#/components/schemas/ReadAloud"
          }
        },
        "type": "object",
        "required": [
          "case",
          "deceased",
          "notes",
          "next_step",
          "read_aloud"
        ],
        "title": "CaseResponse",
        "description": "Opening a case. For a draft this is the resume turn (UC-CASE-10) and counts as activity."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Auth0 access token for the Cairn API audience.",
        "scheme": "bearer"
      }
    }
  },
...
}